feat(search): 接入向量检索与模型重排
- 持久化文档 embedding 并支持旧数据自动补建索引
- 接入 SiliconFlow 和本地兼容 HTTP provider
- 更新检索配置、版本号和双语文档

+++ src/domain.rs
@@ -147,6 +147,6 @@ pub struct SearchResult {
     pub id: Uuid,
     pub title: String,
     pub excerpt: String,
-    pub score: usize,
+    pub score: f32,
     pub updated_at: DateTime<Utc>,
 }
